Remove not used Broadcast and SyncPeer
