+++ protocol/status.go
@@ -1,0 +1,51 @@
+package protocol
+
+import (
+	"math/big"
+
+	"github.com/dogechain-lab/dogechain/protocol/proto"
+	"github.com/dogechain-lab/dogechain/types"
+)
+
+// Status defines the up to date information regarding the peer
+type Status struct {
+	Difficulty *big.Int   // Current difficulty
+	Hash       types.Hash // Latest block hash
+	Number     uint64     // Latest block number
+}
+
+// Copy creates a copy of the status
+func (s *Status) Copy() *Status {
+	ss := new(Status)
+	ss.Hash = s.Hash
+	ss.Number = s.Number
+	ss.Difficulty = new(big.Int).Set(s.Difficulty)
+
+	return ss
+}
+
+// toProto converts a Status object to a proto.V1Status
+func (s *Status) toProto() *proto.V1Status {
+	return &proto.V1Status{
+		Number:     s.Number,
+		Hash:       s.Hash.String(),
+		Difficulty: s.Difficulty.String(),
+	}
+}
+
+// statusFromProto extracts a Status object from a passed in proto.V1Status
+func statusFromProto(p *proto.V1Status) (*Status, error) {
+	s := &Status{
+		Hash:   types.StringToHash(p.Hash),
+		Number: p.Number,
+	}
+
+	diff, ok := new(big.Int).SetString(p.Difficulty, 10)
+	if !ok {
+		return nil, ErrDecodeDifficulty
+	}
+
+	s.Difficulty = diff
+
+	return s, nil
+}
